completed  link route data

+++ TourismSmartTransportation.Data/Models/BasePriceOfBusService.cs
@@ -13,8 +13,8 @@ namespace TourismSmartTransportation.Data.Models
         }
 
         public Guid BasePriceOfBusServiceId { get; set; }
-        public int MaxDistance { get; set; }
-        public int MinDistance { get; set; }
+        public decimal MaxDistance { get; set; }
+        public decimal MinDistance { get; set; }
         public decimal Price { get; set; }
         public int Status { get; set; }
